Remove cmake option for testing (for now). Move testing to the module files themselves.
Tests are generally fast enough to not make very much difference in compilation speeds. The vast majority of the time taken when building the tests was during link-time, so move them to their respective modules.
WIP: Still, the checks in testing should be optional. They are most helpful during development and by providing examples of usage.